+++ app.js
@@ -2,18 +2,16 @@
 require('dotenv').config()
 var express = require('express')
 var path = require('path')
-const bcrypt = require('bcrypt')
 var cookieParser = require('cookie-parser')
 var logger = require('morgan')
-var passport = require('passport')
-var LocalStrategy = require('passport-local').Strategy
+var passport = require('./auth')
 var FacebookStrategy = require('passport-facebook').Strategy
 var session = require('express-session')
 
 var indexRouter = require('./routes/index')
 var accountRouter = require('./routes/account') // automatically looks for index.js in the specified directory
 
-var { findUserByUsernameStrategy, findUserByIdStrategy, findOrCreateUser } = require('./db')
+var { findOrCreateUser } = require('./db')
 
 var app = express()
 
@@ -40,84 +38,16 @@ app.use(express.static(path.join(__dirname, 'public')))
 app.use(passport.initialize())
 app.use(passport.session())
 
-// catch 404 and forward to error handler
-// app.use(function (req, res, next) {
-//   next(createError(404))
-// })
-
 // error handler
 app.use(function (err, req, res, next) {
-  // set locals, only providing error in development
   res.locals.message = err.message
   res.locals.error = req.app.get('env') === 'development' ? err : {}
-
-  // render the error page
   res.status(err.status || 500)
   res.render('error')
 })
 
 app.use('/account', accountRouter)
 app.use('/', indexRouter)
-
-// passport configuration
-passport.use(new LocalStrategy(
-  function (username, password, done) {
-    console.log('🔸 Checking to see if there is a user with that username...')
-    findUserByUsernameStrategy(username)
-      .then(function (result) {
-        if (result) {
-          console.log('✅ User found! This is the user: ', result)
-          const user = result
-          console.log('🔸 Comparing the passwords: ' + user.password + ' 🆚 ' + password)
-          bcrypt.compare(password, user.password)
-            .then(function (bool) {
-              if (user && bool) {
-                console.log('Password matched! ✨')
-                return done(null, user)
-              } else {
-                console.log('Password did not match! 🤮')
-                return done(null, false)
-              }
-            })
-        } else {
-          console.log('Theres no user with that login')
-          return done(null, false)
-        }
-      })
-      .catch(function (error) {
-        console.log('findUserByUsernameStrategy error: ', error)
-        return done(error)
-      })
-  }
-))
-
-passport.serializeUser(function (user, done) {
-  console.log('👉🏻 SERIALIZING')
-  done(null, user)
-})
-
-passport.deserializeUser(function (user, done) {
-  console.log('👉🏻 DESERIALIZING')
-  findUserByIdStrategy(user.userId)
-    .then(function (user) {
-      done(null, user[0])
-    })
-    .catch(function (error) {
-      console.log('deserializeUser err:', error)
-      done(error, null)
-    })
-})
-
-// local strategy
-// passport.use(new Strategy(
-//   function (username, password, cb) {
-//     db.users.findByUsername(username, function (err, user) {
-//       if (err) { return cb(err) }
-//       if (!user) { return cb(null, false) }
-//       if (user.password !== password) { return cb(null, false) }
-//       return cb(null, user)
-//     })
-//   }))
 
 // facebook strategy
 passport.use(new FacebookStrategy({
@@ -127,71 +57,20 @@ passport.use(new FacebookStrategy({
 },
 function (accessToken, refreshToken, profile, done) {
   findOrCreateUser(profile)
-
   done(null, profile)
 }))
 
-// { id: '10157801223336967',
-//   username: undefined,
-//   displayName: 'Alvin Ng',
-//   name:
-//    { familyName: undefined,
-//      givenName: undefined,
-//      middleName: undefined },
-//   gender: undefined,
-//   profileUrl: undefined,
-//   provider: 'facebook',
-//   _raw: '{"name":"Alvin Ng","id":"10157801223336967"}',
-//   _json: { name: 'Alvin Ng', id: '10157801223336967' } }
-
-//   findOrCreateUser({ facebookId: profile.id }, function (err, user) {
-//     if (err) {
-//       return done(err)
-//     } else {
-//       done(null, user)
-//     }
-//   })
-// }
-// ))
-
-// // this is the session id associated with the id
-// passport.serializeUser(function (user, cb) {
-//   cb(null, user)
-// })
-
-// passport.deserializeUser(function (obj, cb) {
-//   cb(null, obj)
-// })
-
-// facebook
 app.get('/auth/facebook',
   passport.authenticate('facebook'))
 
 app.get('/auth/facebook/callback',
   passport.authenticate('facebook', { failureRedirect: '/error' }),
   function (req, res) {
-    res.redirect('/success')
+    res.redirect('/')
   })
 
 app.get('/success', (req, res) => res.send('You have successfully logged in'))
-
 app.get('/error', (req, res) => res.send('error logging in'))
-
-// // catch 404 and forward to error handler
-// app.use(function (req, res, next) {
-//   next(createError(404))
-// })
-
-// // error handler
-// app.use(function (err, req, res, next) {
-//   // set locals, only providing error in development
-//   res.locals.message = err.message
-//   res.locals.error = req.app.get('env') === 'development' ? err : {}
-
-//   // render the error page
-//   res.status(err.status || 500)
-//   res.render('error')
-// })
 
 // listen on port
 app.listen(port, function () {
