+++ db/index.js
@@ -66,7 +66,7 @@ function getAllPosts () {
 }
 
 function findOrCreateUser (user) {
-  console.log('⭐️ Finding or creating a user with Facebook ID of: ', user)
+  console.log('⭐️ Finding or creating a user with Facebook info of: ', user)
   return knex('User').insert({
     fullName: user.displayName
   })
